refactor(messaging-redis): validate StreamEntriesCount and ConnectionPoolSize as int
Replace the uint options with a zero-sentinel post-configure default with
int properties that default to 10 and are validated as greater than zero,
matching the Kafka and NATS transport option shape.

+++ src/Headless.Messaging.Redis/IRedisStreamManagerDefault.cs
@@ -160,12 +160,7 @@ internal sealed class RedisStreamManager(
             var groupedPositions = createdPositions
                 .GroupBy(s => _redis.GetHashSlot(s.Key))
                 .Select(group =>
-                    database.StreamReadGroupAsync(
-                        [.. group],
-                        consumerGroup,
-                        consumerGroup,
-                        (int)_options.StreamEntriesCount
-                    )
+                    database.StreamReadGroupAsync([.. group], consumerGroup, consumerGroup, _options.StreamEntriesCount)
                 );
 
             var readSet = await Task.WhenAll(groupedPositions).ConfigureAwait(false);
